add gtest

+++ include/Node.h
@@ -36,9 +36,9 @@ class Node
   NodeType& get_type() { return _type; }
   NodeState& get_state() { return _state; }
   Node* get_parent_node() { return _parent_node; }
-  int get_known_cost() const { return _known_cost; }
-  int get_est_cost() const { return _est_cost; }
-  int get_total_cost() const { return (_known_cost + _est_cost); }
+  double get_known_cost() const { return _known_cost; }
+  double get_est_cost() const { return _est_cost; }
+  double get_total_cost() const { return (_known_cost + _est_cost); }
   bool get_on_path() { return _on_path; }
 
   void set_coord(const int x,const int y)
@@ -49,8 +49,8 @@ class Node
   void set_type(const NodeType& type) { _type = type; }
   void set_state(const NodeState& state) { _state = state; }
   void set_parent_node(Node* parent_node) { _parent_node = parent_node; }
-  void set_known_cost(const int known_cost) { _known_cost = known_cost; }
-  void set_est_cost(const int est_cost) { _est_cost = est_cost; }
+  void set_known_cost(const double known_cost) { _known_cost = known_cost; }
+  void set_est_cost(const double est_cost) { _est_cost = est_cost; }
   void set_on_path(const bool on_path) { _on_path = on_path; }
   // function
   bool isStart() { return _type == NodeType::kStart; }
@@ -64,8 +64,8 @@ class Node
   NodeType _type = NodeType::kNone;
   NodeState _state = NodeState::kNone;
   Node* _parent_node = nullptr;
-  int _known_cost = 0;
-  int _est_cost = 0;
+  double _known_cost = 0;
+  double _est_cost = 0;
   bool _on_path = false;
 };
